pyproject.toml
[tool.poetry]
name = "ahf-h2h"
version = "0.1.0"
description = "Asian federation of Hand-to-hand combat"
authors = ["Asadbek <ahikmatullayev024@gmail.com>"]
readme = "README.md"
packages = [{include = "ahf_h2h"}]

[tool.poetry.dependencies]
python = "^3.11"
django-debug-toolbar = "^4.4.6"
django-filter = "23.3"
django-environ = "^0.11.2"
psycopg = { version = "*", extras = ["binary", "pool"] }
django-jazzmin = "2.6.0"
pillow = "^10.4.0"
django-rosetta = "^0.10.0"
django-parler = "^2.3"

[tool.poetry.group.dev.dependencies]
Django = "5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
